Drop index only when it's exist

+++ backend/src/graphDB_dataAccess.py
@@ -326,14 +326,14 @@ class graphDBdataAccess:
         param = {"rows":nodes_list}
         return self.execute_query(query,param)
     
-    def drop_create_vector_index(self):
+    def drop_create_vector_index(self, is_vector_index_recreate):
         """
         drop and create the vector index when vector index dimesion are different.
         """
-        self.graph.query("""drop index vector""")
-        
         embedding_model = os.getenv('EMBEDDING_MODEL')
         embeddings, dimension = load_embedding_model(embedding_model)
+        if is_vector_index_recreate == 'true':
+            self.graph.query("""drop index vector""")
         
         self.graph.query("""CREATE VECTOR INDEX `vector` if not exists for (c:Chunk) on (c.embedding)
                             OPTIONS {indexConfig: {
